+++ lib/rivium_push.dart
@@ -2,6 +2,7 @@ library rivium_push;
 
 export 'src/rivium_push.dart';
 export 'src/rivium_push_config.dart';
+export 'src/rivium_push_version.dart';
 export 'src/rivium_push_message.dart';
 export 'src/rivium_push_error.dart';
 export 'src/inapp_message.dart';